requirementos.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
PyMySQL==1.1.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
email_validator


FLASK_ENV=development
SECRET_KEY=troque_esta_chave_por_uma_segura
DATABASE_URL=mysql+pymysql://user:password@localhost:3306/restaurante_db

